Complete MemoryMappedFile class and Junit tests
Todo: find out why AccessDenied exception is thrown when trying to
delete files @After in Junit. The file channel is closed, the
mappedbytebuffer reference is set to null, even the garbage collector
was invoked, but I still get the error, so I have to manually clean the
directory after each run.