Correct totalWeight calculation to include gramms as well. All test passed, problem solved, ticket can be closed.
Done for commit
- Compete calculateSingleCarpetWeight method to pass the test
- Create empty calculateTotalWeight method
- Create Junit test for calculateTotalWeight (expected result = Horton's result)
- Run debug to identify where the calculation differs from Horton's solution
- Complete method calculateTotalWeight to pass the test

Root cause
1. The area calculation was different. My method used integers based on millimeters, Horton's method used doubles based on meters. I have created a new method to calculate area based on meters and debug and Junit test confirmed it yields the same results
2. The totalWeigh calculation was wrong, because when a singleWeigh object was passed to multiply by the total number of carpets, only the kilograms were calculated and gramms were omitted
I created two solutions for this, one is a quick conversion of the singleWeight object to grams, then multiply by total numbers and based on this create the tkg object.
The other solution did exactly the same, because there was  a helper method .toGram() in the tkg class already created. Horton used the same method in his code.